+++ states_rebuilder_package/lib/states_rebuilder.dart
@@ -23,7 +23,7 @@ export 'src/injector.dart' show Injector;
 export 'src/inject.dart' show Inject, Injectable, JoinSingleton;
 
 /// [ReactiveModel] adds a reactive environnement to the injected models.
-export 'src/reactive_model.dart' show ReactiveModel;
+export 'src/reactive_model.dart' show ReactiveModel, RM;
 export 'src/when_connection_state.dart' show WhenRebuilder;
 export 'src/when_rebuilder_or.dart' show WhenRebuilderOr;
 export 'src/on_set_state_listener.dart' show OnSetStateListener;