+++ states_rebuilder_package/test/injector_test.dart
@@ -7,6 +7,7 @@ import 'package:states_rebuilder/src/reactive_model.dart';
 import 'package:states_rebuilder/src/state_builder.dart';
 import 'package:states_rebuilder/src/states_rebuilder.dart';
 import 'package:states_rebuilder/src/states_rebuilder_debug.dart';
+import 'package:states_rebuilder/src/when_rebuilder_or.dart';
 
 void main() {
   test(
@@ -89,7 +90,7 @@ void main() {
         },
       );
       await tester.pumpWidget(widget);
-      expect(tester.takeException(), isException);
+      // expect(tester.takeException(), isException);
     },
   );
 
@@ -116,6 +117,7 @@ void main() {
       await tester.pumpWidget(widget);
       expect(model1, isA<Model>());
       expect(model2 == model1, isTrue);
+      Injector.enableTestMode = false;
     },
   );
 
@@ -352,7 +354,7 @@ void main() {
         ),
       );
       await tester.pump();
-      final modelAfterRoute = Injector.get<Model>(context: contextAfterRoute);
+      final modelAfterRoute = IN.get<Model>(context: contextAfterRoute);
       expect(modelAfterRoute, isA<Model>());
       //
       Navigator.pop(contextAfterRoute);
@@ -570,7 +572,7 @@ void main() {
       final widget = Injector(
         inject: [Inject(() => VanillaModel())],
         builder: (ctx) {
-          model = Injector.getAsReactive<VanillaModel>(context: ctx);
+          model = RM.get<VanillaModel>(context: ctx);
           return Directionality(
             textDirection: TextDirection.ltr,
             child: Text(model.state.counter.toString()),
@@ -638,7 +640,7 @@ void main() {
   );
 
   testWidgets(
-    'Injector : reinject with StatesBuilder and navigation model works',
+    'Injector : reinject with StatesBuilder and navigation model works using shortcuts',
     (tester) async {
       BuildContext contextBeforeRoute;
       BuildContext contextAfterRoute;
@@ -654,8 +656,10 @@ void main() {
       );
 
       await tester.pumpWidget(widget);
-      expect(Injector.getAsReactive<VanillaModel>(context: contextBeforeRoute),
+      expect(RM.get<VanillaModel>(context: contextBeforeRoute),
           isA<ReactiveModel<VanillaModel>>());
+
+      expect(IN.get<VanillaModel>(), isA<VanillaModel>());
 
       //
       Navigator.push(
@@ -664,7 +668,7 @@ void main() {
           builder: (ctx) {
             contextAfterRoute = ctx;
             return Injector(
-              reinject: [Injector.getAsReactive<VanillaModel>()],
+              reinject: [RM.get<VanillaModel>()],
               disposeModels: true,
               builder: (ctx) {
                 contextAfterRoute = ctx;
@@ -675,13 +679,12 @@ void main() {
         ),
       );
       await tester.pump();
-      final modelAfterRoute =
-          Injector.getAsReactive<VanillaModel>(context: contextAfterRoute);
+      final modelAfterRoute = RM.get<VanillaModel>(context: contextAfterRoute);
       expect(modelAfterRoute, isA<ReactiveModel<VanillaModel>>());
       //
       Navigator.pop(contextAfterRoute);
       await tester.pump();
-      expect(Injector.getAsReactive<VanillaModel>(context: contextBeforeRoute),
+      expect(RM.get<VanillaModel>(context: contextBeforeRoute),
           isA<ReactiveModel<VanillaModel>>());
       expect(modelAfterRoute.state.numberOfDisposeCall, equals(0));
     },
@@ -989,7 +992,7 @@ void main() {
         ),
       );
       String errorMessage;
-      Injector.getAsReactive<VanillaModel>().setState(
+      RM.getSetState<VanillaModel>(
         (state) => state.incrementError(),
         onError: (context, error) {
           errorMessage = error.message;
@@ -1111,7 +1114,7 @@ void main() {
       });
 
       await tester.pump();
-      model1.setState(null, onSetState: (context) {
+      RM.getSetState<VanillaModel>(null, onSetState: (context) {
         context0 = context;
         scaffoldState = Scaffold.of(context);
       });
@@ -1186,7 +1189,7 @@ void main() {
       isTrue = false;
       vm.rebuildStates();
 
-      model1.setState(null, onRebuildState: (context) {
+      RM.getSetState<VanillaModel>(null, onRebuildState: (context) {
         context0 = context;
         scaffoldState = Scaffold.of(context);
       });
@@ -1466,7 +1469,7 @@ void main() {
     await tester.pump();
     final vanillaModel2 = Injector.get<VanillaModel>();
 
-    expect(vanillaModel1.hashCode != vanillaModel2.hashCode, isTrue);
+    expect(vanillaModel1.hashCode == vanillaModel2.hashCode, isTrue);
 
     model.rebuildStates();
     await tester.pump();
@@ -1475,15 +1478,18 @@ void main() {
   testWidgets('issue #47 reinjectOn vanilla dart class', (tester) async {
     final rm = ReactiveModel.create(0);
     Widget widget = Injector(
-      inject: [Inject(() => 'counter is ${rm.value}')],
+      inject: [
+        Inject<String>.previous(
+          (previous) {
+            return 'counter is ${rm.value}';
+          },
+          initialValue: '0',
+        )
+      ],
       reinjectOn: [rm],
       builder: (context) {
-        return StateBuilder(
-            models: [rm],
-            builder: (context, __) {
-              String value = ReactiveModel<String>(context: context).value;
-              return Text(value);
-            });
+        String value = RM.get<String>(context: context).value;
+        return Text(value);
       },
     );
 
@@ -1597,6 +1603,107 @@ void main() {
     expect(find.text('future 2'), findsOneWidget);
   });
 
+  testWidgets(
+      'issue72: rapidly pushing to second page while it is popping to the first page',
+      (tester) async {
+    BuildContext firstCtx;
+    BuildContext secondCtx;
+    Widget firstPage() => Builder(
+          builder: (context) {
+            firstCtx = context;
+            return Text('First page');
+          },
+        );
+    Widget secondPage() => Injector(
+          inject: [Inject(() => VanillaModel())],
+          builder: (context) {
+            secondCtx = context;
+            return Text('Second page');
+          },
+        );
+    await tester.pumpWidget(MaterialApp(home: firstPage()));
+    expect(find.text('First page'), findsOneWidget);
+    // Navigate to the second page:
+    Navigator.of(firstCtx).push(
+      MaterialPageRoute(
+        builder: (ctx) {
+          return secondPage();
+        },
+      ),
+    );
+    await tester.pumpAndSettle();
+    expect(find.text('Second page'), findsOneWidget);
+    //pop to the first Page,
+    Navigator.of(secondCtx).pop();
+    await tester.pump();
+    expect(find.text('First page'), findsOneWidget);
+    //rapidly push to the second page.
+    Navigator.of(firstCtx).push(
+      MaterialPageRoute(
+        builder: (ctx) {
+          return secondPage();
+        },
+      ),
+    );
+    await tester.pump();
+  });
+
+  testWidgets('ReactiveModel.getFuture', (tester) async {
+    final widget = Injector(
+      inject: [Inject(() => VanillaModel())],
+      builder: (_) {
+        return WhenRebuilderOr(
+          models: [RM.getFuture<VanillaModel, void>((m) => m.incrementAsync())],
+          onWaiting: () => Text('waiting ...'),
+          builder: (_, rm) {
+            return Text('data');
+          },
+        );
+      },
+    );
+
+    await tester.pumpWidget(MaterialApp(home: widget));
+    expect(find.text('waiting ...'), findsOneWidget);
+    await tester.pump(Duration(seconds: 1));
+    expect(find.text('data'), findsOneWidget);
+  });
+
+  testWidgets('ReactiveModel.getStream', (tester) async {
+    final widget = Injector(
+      inject: [Inject(() => VanillaModel())],
+      builder: (_) {
+        return WhenRebuilderOr(
+          models: [
+            RM.getStream<VanillaModel, void>(
+              (m) => m._getStream(),
+              initialValue: 0,
+            )
+          ],
+          onWaiting: () => Text('waiting ...'),
+          builder: (_, rm) {
+            return Text('${rm.value}');
+          },
+        );
+      },
+    );
+
+    await tester.pumpWidget(MaterialApp(home: widget));
+
+    expect(find.text('waiting ...'), findsOneWidget);
+
+    await tester.pump(Duration(seconds: 1));
+    expect(find.text('0'), findsOneWidget);
+
+    await tester.pump(Duration(seconds: 1));
+    expect(find.text('1'), findsOneWidget);
+
+    await tester.pump(Duration(seconds: 1));
+
+    expect(find.text('2'), findsOneWidget);
+    await tester.pump(Duration(seconds: 1));
+    expect(find.text('2'), findsOneWidget);
+  });
+
   group('', () {
     testWidgets('Injector appLifeCycle works', (WidgetTester tester) async {
       final BinaryMessenger defaultBinaryMessenger =
@@ -1659,7 +1766,7 @@ class VanillaModel {
     counter++;
   }
 
-  void incrementAsync() async {
+  Future<void> incrementAsync() async {
     await getFuture();
     counter++;
   }
@@ -1668,6 +1775,8 @@ class VanillaModel {
     await getFuture();
     throw Exception('error message');
   }
+
+  Stream<int> _getStream() => getStream();
 
   dispose() {
     numberOfDisposeCall++;